--- ehalsmer_auth_config.json
@@ -1,4 +1,0 @@
-{
-  "domain": "dev-69nrm8mx.auth0.com",
-  "clientId": "JvlOT0Pm6jCf6Wl1XsB3cMqqyFU526E0"
-}
