--- routers/moonPhases.js
@@ -1,14 +1,0 @@
-// router for /api/moonPhases, returning data fetched by queryMoonPhase
-
-const express = require('express');
-const data = require('../models/queryMoonPhase');
-
-const router = express.Router();
-
-router.get('/', async (req, res) => {
-    const query = await data.queryMoonPhases();
-    res.status(200).json({message: 'Look at all these express moons!', query})
-})
-
-
-module.exports = router;
