--- test/catmemes.test.js
@@ -1,30 +1,0 @@
-'use strict';
-
-// tests for catmemes
-// Generated by serverless-mocha-plugin
-
-const mochaPlugin = require('serverless-mocha-plugin');
-const expect = mochaPlugin.chai.expect;
-let wrapped = mochaPlugin.getWrapper('catmemes', '/lambda/handler.js', 'catMemes');
-const request = require('supertest')
-
-
-describe('catmemes', () => {
-  it('Cat Memes', () => {
-    it('ok', function it(done) {
-      request('https://bv9cpgqr4l.execute-api.us-east-1.amazonaws.com/dev-nisa/cats')
-        .expect(200)
-        .end(function (error, result) {
-          if (error) {
-            return done(error);
-          }
-  
-          expect(result.body.result).to.deep.eq("it works");
-          done();
-        });
-    });
-    return wrapped.run({}).then((response) => {
-      expect(response).to.not.be.empty;
-    });
-  });
-});
